feat(pado): continuous weekly volume scoring to cut leaderboard ties
Weekly Pado leaderboard scores floored volume into $500 buckets, collapsing
every sub-$500 trader onto 0 volume points. 86.8% of the ~2000 weekly traders
shared a score (424 distinct values, largest tie 46). Make volume scoring
continuous (no floor, same scale/coefficients/caps) so close volumes separate:
measured ties drop to 5.9% (1916 distinct), per-trader score change under 1
point, no top-of-board reordering. The deliberate whale/small-capital
dominance tuning of the weights and caps is left untouched.

- chat-server: extract computeWeeklyVolumePoints (3dp-rounded, side-effect-free
  in leaderboard-types) and call it from the weekly aggregator. total_score /
  score_from_volume become REAL; SQLite affinity preserves the fractional value
  on existing rows, so no migration is needed.
- pado frontend: show 2 decimals for weekly scores below 1K so the added
  resolution is visible (ScoreTraderRow.formatScore).
- network-explorer: round totalScore before the settle-pado PG insert. The
  weekly_score_snapshots.total_score column is INTEGER and a fractional bound
  param fails int4in, which would abort the rank-based weekly settlement (the
  field is display/audit metadata; payouts are rank-based).

+++ apps/nasun-website/chat-server/src/leaderboard-store.ts
@@ -240,9 +240,14 @@ export function initLeaderboardStore(config: LeaderboardConfig): void {
     CREATE TABLE IF NOT EXISTS trader_points_weekly (
       week_id TEXT NOT NULL,
       address TEXT NOT NULL,
-      total_score INTEGER NOT NULL DEFAULT 0,
+      -- total_score / score_from_volume are REAL: weekly volume scoring is
+      -- continuous (computeWeeklyVolumePoints, 3dp) to avoid low-end score ties.
+      -- Existing prod rows on the old INTEGER affinity still accept fractional
+      -- values (SQLite NUMERIC affinity preserves a real with a fractional part),
+      -- so no migration is required; this REAL is for fresh-DB clarity.
+      total_score REAL NOT NULL DEFAULT 0,
       score_from_trades INTEGER NOT NULL DEFAULT 0,
-      score_from_volume INTEGER NOT NULL DEFAULT 0,
+      score_from_volume REAL NOT NULL DEFAULT 0,
       score_from_diversity INTEGER NOT NULL DEFAULT 0,
       score_from_pnl INTEGER NOT NULL DEFAULT 0,
       score_from_prediction_pnl INTEGER NOT NULL DEFAULT 0,
